Moved transitions into the TIled editor layer properties
These are parsed (without error checking - learn to just not make mistakes!) when the map is loaded and are of the format:
  `transition_<id>   :     x:y=>x:y:z`

Where <id> is the identifier of the map you want to transition to. The first x,y are the co-ordinates that are in the current layer, on the current map. The second x:y:z are where you will be taken in the new map